pm_dashboard/requirements.txt
streamlit==1.32.0
pandas==2.1.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
crewai==0.11.0
python-multipart==0.0.6
PyYAML==6.0.1